requirements.txt
snowflake-connector-python[pandas]
fastapi
uvicorn
pydantic
python-multipart
streamlit
altair
pandas
requests
python-dotenv
dbt-snowflake
